Allow Tests To Compile

+++ spark/src/test/scala/geotrellis/spark/io/index/zcurve/Z3RangeSpec.scala
@@ -33,7 +33,7 @@ class Z3RangeSpec extends FunSpec with Matchers {
 
       ranges foreach { case (min, max) =>
         for (z <- min to max) {
-          val zobj = new Z3(z)
+          val zobj = new Z3(z.toLong)
           actualSet =  actualSet + Tuple3(zobj.dim(0),zobj.dim(1),zobj.dim(2))
           count += 1
         }
